+++ Source/Plugins/VolumeOps/UI/VolumeOpsWidget.cpp
@@ -246,7 +246,7 @@ void VolumeOpsWidget::on_m_InputFile_textChanged(const QString &text)
 
     H5VoxelReader::Pointer h5Reader = H5VoxelReader::New();
     h5Reader->setFileName(m_InputFile->text().toStdString());
-    int dims[3];
+    int64_t dims[3];
     float spacing[3];
     float origin[3];
     int err = h5Reader->getSizeResolutionOrigin(dims, spacing, origin);
